refactor: Update and delete redundant code
Matplotlib library was updated so a couple of functions started to show warnings for deprecation. The code was changed to produce the same or at least similar results. Also, some parts in seaborn code were changed to comments to not show the console results but the commands for the different options as it can be redundant, or make the notebook really large.

As summary the changes were made for the parts in:

 - Matplotlib

 - Seaborn