feat: Add stratified sampling in scikit
Stratified sampling is useful to keep the distribution
of the data through the train and test datasets. This method
prevent skewness in the data to avoid errors in predictions
and analysis.

For the example is used a dataset found open source thanks
to the book hands on machine learning from Aurelien Geron.

Resolves: #15