feat: Separate Pandas code and create Pandas Utilities
Separate the Pandas code for it to be easier to memorize as it is the library with more code. Leave the easier and more related to data analysis code in intro to pandas, and leaving pandas utilities what is not directly related to data analysis and advanced code. The name of other files were also changed.