feat: Add introduction to Regex in python.
A regular expression is a sequence of characters that specifies a search pattern in text. Usually such patterns are used by string-searching algorithms for find or replace operations on strings, or for input validation. Regular expression techniques are developed in theoretical computer science and formal language theory. Regex is used in python and more important in Pandas, for optimized feature engineering to use in data analysis or machine learning.